content-curation-pipeline/pyproject.toml
[project]
name = "content-curation-pipeline"
version = "0.1.0"
description = "Content curation pipeline for technical content"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Arpan Patel", email = "arpan@skiplevel.ai" },
]
dependencies = [
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "lxml>=4.9.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "praw>=7.7.0",
    "pydantic>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "tavily-python>=0.3.0",
    "tiktoken>=0.5.0",
    "tokenizers>=0.15.0",
    "youtube-transcript-api>=0.6.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.2.0",
    "pytest>=7.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"] 